requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary
pydantic
pydantic-settings
